Followed the parameters, doesn't look right.
The fact that when printing out a node in the terminal,
it shows a nested node inside hints at a possible error
in both my understanding and my code. To get the value
of the node, I had to invoke '@full_list.value.value'.
This meant that the 'value' inside a node is a nested
node. I have to find out what I did wrong by looking at
other people's submissions.